requirements.txt
autopep8==1.5.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
dill==0.2.9
future==0.16.0
gapic-google-cloud-language-v1==0.15.3
google-api-core==1.19.1
google-auth==1.16.1
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-cloud-language==1.3.0
google-gax==0.15.16
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.29.0
httplib2==0.18.1
idna==2.9
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.5
oauth2client==3.0.0
oauthlib==3.1.0
Pillow==7.1.2
ply==3.8
proto-google-cloud-language-v1==0.15.3
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
toml==0.10.1
tweepy==3.8.0
urllib3==1.25.9
wordcloud==1.7.0
